Debug drawPath python/swift function interop
`Point` arrays in python were not being unpacked quite right when we
tried to draw them as paths. Manually copying the points from the
`Point` buffer and creating new `CGPoints` that swift can use seems to
work.

I am unsure of if the issue is float vs. double or if its something
else.